HashTable Progress
First of all, Happy Turkey day everyone.
Decided to post my progress of the HashTable from last night and this
morning.
Display function works and add function works when there's no
collisions. (Saying this because I haven't tested it with collisions
yet)
Along with the the To-Do list from yesterday, I also need to edit the
remove function to better work with the overflow linked list and then I
will test both add and remove functions for collision handling.
I will be working on this more today (Or at least as much as I can) and
as always, if you have any suggestions or options, let me know.